move database into Daemon class
Move the code into lxd/db.go

and have it always open.
